cpus/M68k/MCF523x/mcf523x.ad[b|s]: functions to get and index useful for IRQ programming
These two subprograms return an index to be used with the IRQ arrays or
the vector table, depending on the value of the VTHandler flag.

If VTHandler is True, the index refers to the Exception Vector table.
If VTHandler is False, the index refers to the register array slot of the
INTC0/1 IRQ controllers.