clibrary/Makefile|c_wrappers.ad[b|s]:
We start translating (slowly) some C library functions in Ada. Actually,
as a matter of fact, malloc/free are Ada, but it was due since the beginning.

The target will be to implement everything, but stdio's variadic functions
(that, in the end, are already calling Ada functions for the ultimate
output of characters)

So, now we begin to try translating the basic ctype and errno headers.

For the use of clibrary, look at the very primitive example in the bsp of
the PC-x86 platform.