Makefile[.os.in|fn.in|.tc.in.ut.in]:
We make some robust changes to the build system in the various Makefiles.

Firstly, we remove the Make version detection in the master Makefile.
The builtin variable MAKE_VERSION exists since 1993, and hopefully no one
has such an ancient version anymore. In case please upgrade your GNU Make.

This will let us to delay the use of the sed utility up to the inclusion of
the toolchain make fragment. So, not being and absolute necessary executable
anymore - up to that point -, we remove it from the initial OS part setup,
and we place it in the Makefile.ut.in fragment, which carries the utilities.
Seems appropriate.

Then, we extract the initial OS detection part in a separate new Makefile
fragment, named Makefile.os.in. This will let us to use it in other contexts,
should them happens.

Then, we provide a multiple makefile-inclusion guard conditional in the
three fragments Makefile.os|ut|fn.in, so there are no problems in including
them more then once.

All in one, an eventually foreign Makefile could take gain of these changes,
and include in sequence:

- Makefile.os.in
- Makefile.ut.in (optional)
- Makefile.fn.in

This is a logical scheme that will replicate a well-defined, toolchain-agnostic,
generic context inside a Makefile. The only environment control variable (apart
the standard machine environment) is VERBOSE, that has to be specified externally
anyway.

All those changes will let us to simplify the management, and will be used in
future commits, e.g., inside the RTS Makefile.

NOTE: the Makefile.ut.in is labelled optional since Makefile.fn.in does not use
any features from it, except UPDATETM (only in Windos environment). In case, a
warning will be issued. If you do not use UPDATETM, simply define it as:

UPDATETM := $(REM)

before the inclusion to silence the warning.

Other minor changes and cosmetics was applied with no particular purpose.