rts/Makefile|src/sfp/Makefile.in:
We make heavy changes also to rts/Makefile (and its include fragment in sfp).

First of all, now this Makefile is a first-citizen of the build system.
This means that it is completely independet, and theoretically all the rts
hierarchy could be relocated elsewhere, regardless SweetAda. It has also
a query mechanism like the master Makefile. It needs only to have PATH set,
so it can found utilities, and the path towards other Makefile fragments.

- add switches so it can be run without a parent Makefile
- does not rely on many environment variables because it loads other support
  Makefile fragments on its own
- compute the set of RTS types
- remove the --RTS switch, which ends up in .ali and it's unuseful at best
- remove the -gnatwa switch, which is redundant
- add local include paths to compilation
- add a query target