Makefile[.tc.in]|rts/Makefile:
We make another set of changes in order to make the build system modular
and flexible.

1)
Makefiles now are logically indipendent, and do not require a master Makefile.
Anyway, the master Makefile provides the environment variables in order to carry
on the compilation. But they are not tied anymore to some setup choices and
could even be relocatable.

2)
Makefiles have now a probing mechanism that output multiple variables
NOTE: multiple-variable queries are separated by the "|" character,
and are a fully assignment statement,so that they can be evaluated in
a script

3)
Remove redundant/unuseful variables:
- RTS_BASE_PATH, since is only a temporary to define RTS_PATH
- RTS_ROOT_PATH, the same
- USE_LIBADA, since we link always with libgnat.a/libgnarl.a

4)
Add the GPRCLEAN variable that carries the gprclean executable, and
EXTERNAL_ALIS, which is useful to bind with an external library

5)
Thank to the 2), Makefile can query the RTS Makefile about the libraries
to be included in the final executable, rather than synthesize names
on its own

6)
cosmetics and adjustment