MSP432P401R/msp432p401r.ads: 1) add a _UART suffix to all registers in UART mode 2) encapsulate UCAxBRW and UCAxIV in a record for uniformity 3) the ASYNC/SYNC mode flag is a Bits_1, not a Boolean (so use    UCSYNC_ASYNC/UCSYNC_SYNC values) 4) report exact comments as per datasheet 5) definitions for eUSCI A SPI mode
Note that, in 5), we use the same constants used for UART mode, so some
definitions are commented out (but valid)

I feel that UART/SPI structure will be modified in the future. For now
we take these for granted, maybe only some prefixes in user code entities
will need small changes.