PC-x86/QEMU-ROM/qemu.sh:
Only for this platform, and only for non-cmd environments (so far), enable
CAN device emulation.

The CAN emulation needs "export QEMU_CAN=Y" in the configuration.in file,
and a setup for creating a virtual CAN network in the host. In Linux,
you could do a thing like this:

$ modprobe vcan
$ ip link add dev vcan0 type vcan
$ ip link set vcan0 up

See the reference documentation in the can-utils package for additional
informations.

Plus, rename QEMU_DEBUG to QEMU_DEBUG_OPTIONS.