Makefiletc.in/RTS/Makefile/[configuration.in/rts.cs.in]:
Remove rts.cs.in in the RTS build tree, which contains ADAC/CC RTS compilation
switches. Now switches are contained (and available) in the configuration.in
file, thanks to the query feature of Makefiles. This allow us to centralize
informations.

Thanks to this feature, there is now a complete decoupling of the RTS
build (from a logical layout point of view), and we increase (and simplify)
to a higher degree the RTS configuration and his manipulation, e.g., control
of switches depending on RTS type.