core/memory: new unit
The purpose of the Memory unit is to provide a database of memory segments
to be used by the system (e.g., mmu and/or malloc). In this initial layout,
it must be allocated and initialized -- most likely -- in the startup.S
code, before Ada code run. But it could be (erased and) filled later by the
system, with some kind of resource discovering.

The memory map is an array of memory blocks, where a block is a record
with a start address, a size and a type.

The array has a fixed number of entries, NMEMORYAREAS_MAX, and the startup
code stores in NAreas the real number of slots used. The physical layout
of the memory block is implemented in this way so it should be simple to
define it also in asm language.