Makefile[ck.in|os.in],configuration.in:
We remove the tools check machinery, making it a target. This way Makefile
has not to check every time the presence of the tools if one leaves the
variable TOOLS_CHECK active. When SweetAda is installed, run this target
once in order to check whether tools like sed,gcc|gnat-wrapper, and maybe
elftool, are operational. We remove the now unused variable TOOLS_CHECK
from configuration.in.

Plus, define the COMMA variable together with the other special characters,
and specify well the "probevariables" target in "make help".