RaspberryPi3: first implementation of (simple) multi-core on real hardware
First of all, we align configuration with last rules about locally defining
the RTS and the PROFILE type. We choose SFP since we don't have memory
constraints.

Then, we create a straight memory mapping for MMU. Without MMU we can't run
all the 4 cores in parallel. Neither we can enable cache.

This way we can run all the 4 cores with the same mechanism outlined in the
QEMU-AArch64 example.

Since we haven't (yet) a cache management API, for make things simpler,
for now we allocate a tiny segment of shared RAM in uncached memory.

The 4 core run respectively:

core 0: nothing, waiting for the interrupt timer that will be managed in
        exceptions.adb (and will blink a LED in GPIO05)
core 1: dump CPU parameters on mini-UART: clocks, voltages, temperature
core 2: blink a LED in GPIO06
core 3: continously print an incrementing variable on mini-UART

You need to connect a serial port on mini-UART pins 14/15 to read outputs,
and something to see the LEDs blinking on GPIO05/6.