Makefile.lb.in,cpus/*/cpu.gpr:
We remove the last unstructured construct (read: hack) from the build system.

The problem was, libcpu is the only library that could be empty, this happens
when a CPU does not need low-level support for simple things.
Unfortunately, there are problems when operating under GPRbuild, and so far
we used to circumvent this with a temporary solution, by using an artificial
Makefile variable to be used manually (albeit it was never necessary to do
so, it was just a "feature" for the user who wants to manage its personal
build).

Now, we call unconditionally GPRbuild (which is the standard operation),
and we trigger the existence of source files when at east one should be
included in the library. At the end of the processing, if no files exist,
we disable the compilation by means of a flag. This will prevent errors
from the tool, because an empty set of source files isn't a tolerated
option.

When GPRbuild exits, we check if the library exists. If it doesn't, then we
create an empty library. This will prevent errors from the linker.