Makefile[.fn.in|.lb.in]
After all, putting a space between -I and an include directory was not a
great idea, due to problems in GNATMAKE handling. Revert this.

Plus, now the update-timestamp call is rather convoluted. So directly
execute the provided script.

This, in turn, let us to delete the update-timestamp function from the
Makefile.fn.in, make it independent from the Makefile.ut.in fragment.
So the Makefile.ut.in and the Makefile.fn.in fragments are now includeable
in whatever order we want. They depend only on Makefile.os.in, which is the
main base fragment of every Makefile.