Makefile[.lb.in|.tc.in|.ut.in]:
Build system revision.

We remove the sed utility from windowse environment. And since sed is a
standard POSIX utility, it doesn't need to be exported nor referenced as
a variable.

We also apply some enhancements and shortcuts to make scripts more flexible
and modular.

We also protect the build system from being fooled by random targets issued
in command line.

As a consequence of this, SweetAda is now completely independent from the
underlying OS, and needs no support in terms of 3rd party tools. Just the
toolchain and GNU make (being it the standard tool supplied in Linux,
homebrew's Make for maccOS (or similar), MSYS2's make, cmd.exe's mingw32-make.
And od course GPRbuild, for those users that want that build mode in place
of plain GNATMAKE.

Install GNU Make version 4.4 and later.

SweetAda now is a clean-room, 100% sourced system.

In a short time we will remove the Make and sed zipped executables from
the repository.