rts/src/all targets/configuration.in|zfp/configuration.in|sfp/configuration.in:
We start to gather compiler flags in system.ads that are meant to be constant
across target (CPU type) configurations and RTS types, setting them in the
default-provided zfp|sfp/configuration.in.

Remember that, anyway, the target CPU configuration could override these
values.

This way is a little bit easier to check what the target needs to set
in terms of Ada frontend/backend features.