STM32F769I/stm32f769i.ad[b|s]:
Definitions for NAND Flash controller and creation of unit body.

We remove the loooong list of constants for bitfields with length >= 256.

Instead, we preserve as constants only default values, then, we define
inlined functions to compute the right value, simulating the entire set
of constant for that bitfield. This way, values are checked and the unit
is still preelaborable.