configuration.in|Makefile|libutils/createconfiguregpr.[ps1|sh]:
We introduce the ELABORATION_MODEL variable, which -- for the moment -- has
no effect.

This system-wide variable (overridable, as usual, in platform's configuration.in)
will decide the elaboration model to be used: STATIC, DYNAMIC, SPARK.

We will take gain of it when we will slowly perform a transition to a
more SPARK-friendly codebase, since we will need the SPARK model to run
tools like, e.g., gnatprove.