gnatmake_tc.in|gprbuild_tc.gpr:
Enable elaboration mode selection.

Now you can select the elaboration model by setting the variable in the
global top-level configuration.in (or in the platform's one):

ELABORATION_MODE := <empty> | DYNAMIC | SPARK

<empty> means the default GNAT elaboration model.

Note that SPARK elaboration model is mostly work-in-progress since SweetAda
codebase is pure Ada so far (but it is legal to use it and no errors should
show up in compilation).