core/bits.ad[b|s]:
We make some changes to Bits unit.

1)
cosmetics: remove "in" mode specifier in some functions to align the declaration
style with those of other functions.

2)
shorten names for Bits_1 <=> Boolean conversions, not a great loss of
readability and comprehension:
  To_Bits_1 -> To_B1
  To_Boolean -> To_B

3)
About the types PBoolean and NBoolean, conversions are somewhat convoluted
and may lead to confusion. NBoolean exists because of hardware signals that
have negative logic (e.g., the /RTS signal in RS-232C, or the /BUSY signal
in parallel printer port). They are active-low and are asserted when signals
carry a logic 0. PBoolean is really artificial and redundant, but it is
here just for completeness and could be used to highlight differences.

Instead of Active/Inactive functions, we use the trivial conversions to/from
Booleans, but with different names and semantics: if a signal is declared
NBoolean, you just convert to/from Booleans with a To_B/To_N function, and you
just think with positive logic, because, in the end, "if" Ada statement map
only to Booleans.

Given the following declarations:

SerialShouldDoRTS : Boolean;  -- positive logic
nRTS              : NBoolean; -- negative logic

Then:

-- nRTS will be at logic 0 (asserted) when SerialShouldDoRTS is True
-- nRTS will be at logic 1 (deasserted) when SerialShouldDoRTS is False
nRTS = To_N (SerialShouldDoRTS);

Obviously (the if is a little bit nonsense):

if To_B (nRTS) then
   -- True, nRTS is asserted and /RTS is at logic 0 (active)
else
   -- False, nRTS is not asserted and /RTS is at logic 1 (inactive)
end if;

This way of thinking somewhat resembles also true hardware design, because
the signals internaly are almost always active-high, just the front-end is
a buffer driver that does the inversion.

Not a great deal, but hopefully is more clear and standardized. If you have
already used Active/Inactive functions, just replace:

- Active => To_B ()
- Inactive => not To_B ()