core/core*, core/configuration.in:
We introduce a logical fragmentation in the core unit, one base unit always
used, and one "expanded" unit only for bigger RTSes other than zfp and sfp
(that so far do not exist yet).

Note that RTSes can be configurable in SweetAda, and you can build your own,
provided an RTS configuration that includes the proper files (maybe copied
from GNAT sources).

When using high-level features, say, finalization, you must instruct the
binder to behave in a certain way. This implies setting the flag
Suppress_Standard_Library := False in system.ads. This flag is always False
for zfp and sfp. To activate more features, the flag should be set to True,
and that could be done in the RTS' configuration.in, which will trigger this
thing when creating such an RTS.

But the binder, at this point, needs a set of variable objects that should
be exposed by the system, to satisfy some interbal references, even if not
used.

The package Core.Expanded will satisfy this requirement.