core/core-base.ads: this commit is an artifact: read for clarification
two commit ago we pushed a commit for core/core[base|expanded] for new
features.

Unfortunately, the message was wrong, saying that Suppress_Standard_Library
is normally set to False for zfp/sfp, and should be made True to activate
other features.

It's the converse: Suppress_Standard_Library is normally True for zfp/sfp
and should made False for new, expanded RTSes. Just for clarity.

We do not amend the message because this will force obscure mechanism in Git,
and we prefer to create an artificial commit to inform people.