rts/mfp:
We introduce the "medium footprint profile". This is a 1-1 copy of sfp,
but has finalization (not used anyway). It allows to link in small libraries
or modules that require this feature. Maybe this type of RTS will grow
in the future.

Enable it in platforms/<your_platform>/configuration.in with:
RTS := mfp
PROFILE := mfp

Remember to run "make rts" to generate all RTS units for your cpu.